Refactor: `MainPage` 구조 변경
- 메인 페이지 내부에서 로그인, 회원가입 화면 간 전환을 해당 부분만을 교체하는 방법으로 진행할 예정 -> 코드 수정
- pages 디렉토리 하위 main 디렉토리를 추가로 생성하여 main 페이지 구성에 필요한 컴포넌트 및 페이지를 관리
- 불필요한 import 구문 제거
